refactor: allow async for latency sensitive
It turns out, that while maintaining max throughput, latency is far
from perfect. Moving from a max throughput to a min latency model
the user is prompted much much faster with the first output.

For CI where stdout / stderr are pumped to file, this also
allows for an improved resource pattern.